+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentSheetVerticalViewController.swift
@@ -14,7 +14,7 @@ import UIKit
 class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewControllerProtocol, PaymentSheetViewControllerProtocol {
     enum Error: Swift.Error {
         case missingPaymentMethodListViewController
-        case missingPaymentMethodFormViewController
+        case missingContentViewController
         case noPaymentOptionOnBuyButtonTap
     }
     var selectedPaymentOption: PaymentSheet.PaymentOption? {
@@ -41,15 +41,15 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
             case .saved(paymentMethod: let paymentMethod):
                 return .saved(paymentMethod: paymentMethod, confirmParams: nil)
             }
-        } else {
-            // Otherwise, we must be showing the form - use its payment option
-            guard let paymentMethodFormViewController else {
-                stpAssertionFailure("Expected paymentMethodFormViewController")
-                let errorAnalytic = ErrorAnalytic(event: .unexpectedPaymentSheetError, error: Error.missingPaymentMethodFormViewController)
-                STPAnalyticsClient.sharedClient.log(analytic: errorAnalytic)
-                return nil
-            }
+        } else if let paymentMethodFormViewController {
             return paymentMethodFormViewController.paymentOption
+        } else if isRecollectingCVC, let cvcRecollectionViewController {
+            return cvcRecollectionViewController.paymentOption
+        } else {
+            stpAssertionFailure()
+            let errorAnalytic = ErrorAnalytic(event: .unexpectedPaymentSheetError, error: Error.missingContentViewController, additionalNonPIIParams: ["error_message": "Missing content! Expected list, form, or cvc", "first_child_vc": String(describing: children.first)])
+            STPAnalyticsClient.sharedClient.log(analytic: errorAnalytic)
+            return nil
         }
     }
     // Edge-case, only set to true when Link is selected via wallet in flow controller
@@ -88,13 +88,17 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
         // Edge case: If Apple Pay isn't in the list, show Link as a wallet button and not in the list
         loadResult.isLinkEnabled && isFlowController && shouldShowApplePayInList
     }
-    // Whether or not we are in the special case where we don't show the list and show the card form directly
+    /// Whether or not we are in the special case where we don't show the list and show the card form directly
     var shouldDisplayCardFormOnly: Bool {
         return paymentMethodTypes.count == 1 && paymentMethodTypes[0] == .stripe(.card)
         && savedPaymentMethods.isEmpty
         && !shouldShowApplePayInList
         && !shouldShowLinkInList
     }
+    /// The content offset % of the payment method list before we transitioned away from it
+    var paymentMethodListContentOffsetPercentage: CGFloat?
+    /// True while we are showing the CVC recollection UI (`cvcRecollectionViewController`)
+    var isRecollectingCVC: Bool = false
 
     private lazy var savedPaymentMethodManager: SavedPaymentMethodManager = {
         SavedPaymentMethodManager(configuration: configuration, intent: intent)
@@ -113,6 +117,7 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
 
     var paymentMethodListViewController: VerticalPaymentMethodListViewController?
     var paymentMethodFormViewController: PaymentMethodFormViewController?
+    var cvcRecollectionViewController: CVCReconfirmationVerticalViewController?
 
     lazy var paymentContainerView: DynamicHeightContainerView = {
         DynamicHeightContainerView()
@@ -202,6 +207,9 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
             if let override = paymentMethodFormViewController?.overridePrimaryButtonState {
                 return override.ctaType
             }
+            if isRecollectingCVC {
+                return .custom(title: String.Localized.confirm)
+            }
             if let customCtaLabel = configuration.primaryButtonLabel {
                 return isFlowController ? .custom(title: customCtaLabel) : .customWithLock(title: customCtaLabel)
             }
@@ -214,6 +222,9 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
         let state: ConfirmButton.Status = {
             if isPaymentInFlight {
                 return .processing
+            }
+            if let cvcRecollectionViewController, isRecollectingCVC {
+                return cvcRecollectionViewController.paymentOptionIntentConfirmParams == nil ? .disabled : .enabled
             }
             if let override = paymentMethodFormViewController?.overridePrimaryButtonState {
                 return override.enabled ? .enabled : .disabled
@@ -273,9 +284,19 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
 
     func updateError() {
         errorLabel.text = error?.nonGenericDescription
-        animateHeightChange {
-            self.errorLabel.setHiddenIfNecessary(self.error == nil)
-        }
+        animateHeightChange({ [self] in
+            errorLabel.setHiddenIfNecessary(error == nil)
+            if error != nil {
+                // Without this, setting `contentOffsetPercentage = 1` uses the height of the scroll view without accounting for the error label
+                errorLabel.setNeedsLayout()
+                errorLabel.layoutIfNeeded()
+            }
+        }, postLayoutAnimations: {
+            if self.error != nil {
+                // Scroll the view to the bottom to ensure the error is visible
+                self.bottomSheetController?.contentOffsetPercentage = 1
+            }
+        })
     }
 
     func makePaymentMethodListViewController(selection: VerticalPaymentMethodListSelection?) -> VerticalPaymentMethodListViewController {
@@ -472,6 +493,14 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
                     #if !canImport(CompositorServices)
                     UINotificationFeedbackGenerator().notificationOccurred(.error)
                     #endif
+
+                    let nsError = error as NSError
+                    let isCVCError = nsError.domain == STPError.stripeDomain && nsError.userInfo[STPError.errorParameterKey] as? String == "cvc"
+                    if isRecollectingCVC, !isCVCError {
+                        // If we're recollecting CVC, pop back to the main list unless the error is for the cvc field
+                        sheetNavigationBarDidBack(navigationBar)
+                    }
+
                     // Update state
                     self.isUserInteractionEnabled = true
                     self.error = error
@@ -521,6 +550,25 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
             stpAssertionFailure("Tapped buy button while adding without paymentOption")
             return
         }
+
+        // If the selected payment option is a saved card, CVC is enabled, and we are PS, handle CVC specially:
+        if case let .saved(paymentMethod, _) = selectedPaymentOption, paymentMethod.type == .card, intent.cvcRecollectionEnabled, !isFlowController, !isRecollectingCVC {
+            let cvcRecollectionViewController = CVCReconfirmationVerticalViewController(
+                paymentMethod: paymentMethod,
+                intent: intent,
+                configuration: configuration,
+                elementDelegate: self
+            )
+            self.cvcRecollectionViewController = cvcRecollectionViewController
+            isRecollectingCVC = true
+            paymentMethodListContentOffsetPercentage = bottomSheetController?.contentOffsetPercentage
+            switchContentIfNecessary(to: cvcRecollectionViewController, containerView: paymentContainerView)
+            navigationBar.setStyle(.back(showAdditionalButton: false))
+            error = nil
+            updateUI()
+            return
+        }
+
         pay(with: selectedPaymentOption)
     }
 
@@ -568,10 +616,6 @@ extension PaymentSheetVerticalViewController: BottomSheetContentViewController {
 
     var requiresFullScreen: Bool {
         return false
-    }
-
-    func didFinishAnimatingHeight() {
-        // no-op
     }
 }
 
@@ -630,7 +674,8 @@ extension PaymentSheetVerticalViewController: VerticalPaymentMethodListViewContr
             if pmFormVC.form.collectsUserInput {
                 // The payment method form collects user input, display it
                 self.paymentMethodFormViewController = pmFormVC
-                switchContentIfNecessary(to: pmFormVC, containerView: paymentContainerView, resettingContentOffset: true)
+                paymentMethodListContentOffsetPercentage = bottomSheetController?.contentOffsetPercentage
+                switchContentIfNecessary(to: pmFormVC, containerView: paymentContainerView, contentOffsetPercentage: 0)
                 navigationBar.setStyle(.back(showAdditionalButton: false))
             }
         }
@@ -695,6 +740,7 @@ extension PaymentSheetVerticalViewController: SheetNavigationBarDelegate {
         view.endEditing(true)
         error = nil
         paymentMethodFormViewController = nil
+        cvcRecollectionViewController = nil
         guard let paymentMethodListViewController else {
             stpAssertionFailure("Expected paymentMethodListViewController")
             let errorAnalytic = ErrorAnalytic(event: .unexpectedPaymentSheetError, error: Error.missingPaymentMethodListViewController)
@@ -702,8 +748,11 @@ extension PaymentSheetVerticalViewController: SheetNavigationBarDelegate {
             didCancel()
             return
         }
-        paymentMethodListViewController.clearSelection()
-        switchContentIfNecessary(to: paymentMethodListViewController, containerView: paymentContainerView)
+        if !isRecollectingCVC {
+            paymentMethodListViewController.clearSelection()
+        }
+        isRecollectingCVC = false
+        switchContentIfNecessary(to: paymentMethodListViewController, containerView: paymentContainerView, contentOffsetPercentage: paymentMethodListContentOffsetPercentage)
         navigationBar.setStyle(.close(showAdditionalButton: false))
         updateUI()
     }
@@ -740,6 +789,7 @@ extension PaymentSheetVerticalViewController: UpdateCardViewControllerDelegate {
 
 extension PaymentSheetVerticalViewController: PaymentMethodFormViewControllerDelegate {
     func didUpdate(_ viewController: PaymentMethodFormViewController) {
+        error = nil  // clear error
         updateUI()
     }
 
@@ -763,5 +813,18 @@ extension PaymentSheetVerticalViewController: WalletHeaderViewDelegate {
         }
 
         paymentSheetDelegate?.paymentSheetViewControllerDidSelectPayWithLink(self)
+    }
+}
+
+// MARK: - ElementDelegate
+/// Used for CVC Recollection - we are the delegate of the CVC element
+extension PaymentSheetVerticalViewController: ElementDelegate {
+    func continueToNextField(element: Element) {
+        updateUI()
+    }
+
+    func didUpdate(element: Element) {
+        self.error = nil
+        updateUI()
     }
 }